caeml: add -H flag to specify headers
Add -H flag to specify headers as a comma-separated list. The order of
the header values is determined by the order in the list. Any header
value can be displayed. For example,

caeml -H "From,To,Cc,Bcc,Date,Subject"

corresponds to the default case if caeml is called without any
arguments. However, we could also do the following now:

caeml -H "Subject,To,From,Message-Id"

Signed-off-by: Koni Marti <koni.marti@gmail.com>